+++ Driver/FileOperate.h
@@ -32,8 +32,8 @@ u8 File_Mkdir(const char * Path);
 
 //File Operate   ÎÄ¼þ²Ù×÷
 u8 File_CreateNewFile(const char* Path);
-u8 File_ReadData(const char* Path,u8* Data,u16 Length,u32 Offset);
-u8 File_WriteData(const char* Path,u8* Data,u16 Length,u32 Offset);
+u8 File_ReadData(const char* Path,u8* Data,u32 Length,u32 Offset);
+u8 File_WriteData(const char* Path,u8* Data,u32 Length,u32 Offset);
 u32 File_GetFileSize(const char* Path);
 void File_GetFileNameList(void);
 u8 File_Rename(const char* path_old, const char* path_new);
